Move elements to components in Uttaksplan
