+++ src/uttaksplan/components/ikoner/InfoIkon.tsx
