+++ src/uttaksplan/components/ikoner/LukkInfoIkon.tsx
