+++ src/uttaksplan/components/ikoner/PlussIkon.tsx
