+++ src/uttaksplan/components/ikoner/TerminIkon.tsx
