+++ src/uttaksplan/components/utvidetInformasjon/InfoToggler.tsx
