+++ src/uttaksplan/components/utvidetInformasjon/infoToggler.less
