+++ src/uttaksplan/components/utvidetInformasjon/utvidetInformasjon.less
